MLOps/requirements.txt
apache-airflow==2.5.1
apache-airflow-providers-python==3.1.0
pandas==1.5.3
scikit-learn==1.2.2
mlflow==2.3.1
numpy==1.24.2
joblib==1.2.0